requirements-dev.txt
Flask
SQLAlchemy
psycopg2-binary
python-dotenv
bcrypt

pytest
pytest-cov
pytest-mock

coverage
